Fixed the header
I added hover effect for the footer links
I made the logo a link. There was a problem with the phone-icon when I did that but got it fixed. Also I had to add the font-family to the a link as it did not get it from the logo id
I worked on the icons in the header as they were misplacing the whole page. I managed to sort them by changing their properties. I removed the positioning and added some margins instead I also added float: right;.