Products page update
I added new photos for the products. I downloaded photos and then cropped and resized them to look the same on the website.
I fixed some small issues here and there on the products page
I started to work on its responsiveness and had some issues with the order button. It gets misplaced when the screen size changes.
There was an issue with the flex display property as it allowed the cards to move as much as they wanted on screen size change. I managed to get that fixed by adding max-width to the cards container and auto margin. Now they only change their display on the breakpoints
I fixed my code and added comments