IE fixes
I managed to fix the background-color issue for <hr class="rule"> by adding a container div around it. I styled it with white border and background-color and it shows correctly now.
I fixed an issue on IE where the text in staff images was getting to the very edge of the page. I fixed it. It was an issue on small screen sizes only
 The drop-down in the header on small screens did not show the items. The problem was that they were left aligned. I added code to align them to the right of the page
The staff images get on top of each other when the screen get too big (for example 2 screens - over 2500px). I added a max width and height to avoid that.