Small screen fixes for home page
I fixed the responsiveness issue for the icon-headlines. I made them a bit smaller and removed the padding in order to fit on the page
I changed the padding for the text in the article as it was following the big screen one and it was too big and no much room for the text itself on small screen
I fixed the .circle-splash as it was not displaying correctly on small screens. I added an extra breakpoint as it was missing previously
The owl-carousel is also not looking good on small screens. I spent some time to see how the code was written in the first place as it seems there are a lot of dependencies. I change the image to be smaller (scale(0.7)). I also changed the arrows' position to underneath the image.
I also made the buttons in product carousel and reviews one be smaller on small screen sizes
I fixed the distance between the dots in the reviews carousel for small screens